代码优化，评论系统的开发，部署配置文件的更新

+++ internal/pkg/kit/constant/space.go
@@ -1,6 +1,6 @@
 package constant
 
-type SpaceVisitorType = int
+type SpaceVisitorType = int64
 
 const (
 	SpacePublicVisitor SpaceVisitorType = 1
